requirements.txt
python-telegram-bot[job-queue]==21.7
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
gspread==6.1.2
google-auth==2.35.0
python-dotenv==1.0.1
tzdata==2024.2

SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1
